requirements.txt
google-cloud-speech
google-cloud-texttospeech
huggingface_hub
sounddevice
soundfile
pygame
